==> 3_dl/dl-regression/1.regression.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, LabelEncoder, OneHotEncoder
import pickle

df_main = pd.read_csv('data_Churn_Modelling.csv')
df_main.head()

# Preprocess the data
df_main = df_main.drop(['RowNumber', 'CustomerId', 'Surname'], axis=1)
df_main.head()

# Encode categorical variables
label_encoder_gender = LabelEncoder()
df_main['Gender'] = label_encoder_gender.fit_transform(df_main['Gender'])
df_main.head()

list(df_main['Geography'].unique())

# One-hot encode 'Geography'
onehot_encoder = OneHotEncoder(handle_unknown='ignore')
geo_encoded = onehot_encoder.fit_transform(df_main[['Geography']]).toarray()
geo_columns = onehot_encoder.get_feature_names_out(['Geography'])
df_geo_encoded = pd.DataFrame(geo_encoded, columns=geo_columns)
df_main = pd.concat([df_main.drop('Geography', axis=1), df_geo_encoded], axis=1)
df_main.head()


# Split the data into features and target
X = df_main.drop('EstimatedSalary', axis=1)
y = df_main['EstimatedSalary']

X.columns

## Split the data in training and tetsing sets
X_train,X_test,y_train,y_test = train_test_split(X,y,test_size=0.2,random_state=69)

## Scale these features
scaler=StandardScaler()
X_train=scaler.fit_transform(X_train)
X_test=scaler.transform(X_test)

# Save the encoders and scaler for later use
with open('saved_label_encoder_gender.pkl', 'wb') as file:
    pickle.dump(label_encoder_gender, file)

with open('saved_onehot_encoder_geo.pkl', 'wb') as file:
    pickle.dump(onehot_encoder, file)

with open('saved_scaler.pkl', 'wb') as file:
    pickle.dump(scaler, file)

#### ANN Regression Problem statement

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from tensorflow.keras.callbacks import EarlyStopping, TensorBoard
import datetime

# Build the model
model = Sequential([
    Dense(64, activation='relu', input_shape=(X_train.shape[1],)),
    Dense(32, activation='relu'),
    Dense(1)  # Output layer for regression
])
model.compile(optimizer='adam',loss='mean_absolute_error',metrics=['mae'])
model.summary()

# Set up TensorBoard
log_dir = "logs/fit/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
tensorboard_callback = TensorBoard(log_dir=log_dir, histogram_freq=1)

# Set up Early Stopping
early_stopping_callback = EarlyStopping(monitor='val_loss', patience=10, restore_best_weights=True)

# Train the model
history = model.fit(
    X_train, y_train,
    validation_data=(X_test, y_test),
    epochs=100,
    callbacks=[early_stopping_callback, tensorboard_callback]
)
model.save('saved_model_regression.h5')

%load_ext tensorboard

%tensorboard --logdir logs/fit

## Evaluate model on the test data
test_loss,test_mae = model.evaluate(X_test,y_test)
print(f'Test MAE : {test_mae}')
